refactor: add pytest for tests.   change directory of test   using class for tests to grouping.
